feat: adding tests for all user services but update role
