+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { UsersController } from './users/users.controller';
 import { PrismaService } from './prisma/prisma.service';
 import { UsersService } from './users/users.service';
-import { UsersModule } from './users/tests/users.module';
+import { UsersModule } from './users/users.module';
 import { DirectionsModule } from './directions/directions.module';
 import { DepartmentsModule } from './departments/departments.module';
