+++ src/users/__tests__/users.controller.spec.ts
@@ -1,12 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UsersController } from '../users.controller';
+import { PrismaService } from '../../prisma/prisma.service';
+import { UsersService } from '../users.service';
 
-describe('LoginController', () => {
+describe('UsersController', () => {
   let controller: UsersController;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [UsersController],
+      providers: [PrismaService, UsersService],
     }).compile();
 
     controller = module.get<UsersController>(UsersController);
