+++ src/users/users.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
-import { UsersService } from '../users.service';
+import { UsersService } from './users.service';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { UsersController } from './users.controller';
 
 @Module({
   providers: [UsersService, PrismaService],
   exports: [UsersService],
+  controllers: [UsersController],
 })
 export class UsersModule {}
